+++ src/test/java/github/weichware10/util/data/DataPointTest.java
@@ -15,11 +15,11 @@ public class DataPointTest {
     @Test
     public void dataIdShouldBeSetCorrectly() {
         int id1 = (int) Math.random() * 1000;
-        DataPoint dataPoint1 = new DataPoint(id1, 0, new int[] { 0, 0 }, new int[] { 0, 0 });
+        DataPoint dataPoint1 = new DataPoint(id1, 0, new double[] { 0, 0 }, new double[] { 0, 0 });
         assertEquals("dataPoint1.dataId sollte " + id1 + " sein", id1, dataPoint1.dataId);
 
         int id2 = (int) Math.random() * 1000;
-        DataPoint dataPoint2 = new DataPoint(id1, 0, new int[] { 0, 0 }, 0.0f);
+        DataPoint dataPoint2 = new DataPoint(id1, 0, new double[] { 0, 0 }, 0.0f);
         assertEquals("dataPoint2.dataId sollte " + id2 + " sein", id2, dataPoint2.dataId);
     }
 
@@ -29,12 +29,12 @@ public class DataPointTest {
     @Test
     public void timeOffsetShouldBeSetCorrectly() {
         int id1 = (int) Math.random() * 1000;
-        DataPoint dataPoint1 = new DataPoint(0, id1, new int[] { 0, 0 }, new int[] { 0, 0 });
+        DataPoint dataPoint1 = new DataPoint(0, id1, new double[] { 0, 0 }, new double[] { 0, 0 });
         assertEquals("dataPoint1.timeOffset sollte " + id1 + " sein",
                 id1, dataPoint1.timeOffset, 0.0001f);
 
         int id2 = (int) Math.random() * 1000;
-        DataPoint dataPoint2 = new DataPoint(0, id2, new int[] { 0, 0 }, 0.0f);
+        DataPoint dataPoint2 = new DataPoint(0, id2, new double[] { 0, 0 }, 0.0f);
         assertEquals("dataPoint2.timeOffset sollte " + id2 + " sein",
                 id2, dataPoint2.timeOffset, 0.0001f);
     }
@@ -44,12 +44,14 @@ public class DataPointTest {
      */
     @Test
     public void coordinatesShouldBeSetCorrectly() {
-        int[] coordinates1 = new int[] { (int) Math.random() * 1000, (int) Math.random() * 1000 };
-        DataPoint dataPoint1 = new DataPoint(0, 0, coordinates1, new int[] { 0, 0 });
+        double[] coordinates1 = new double[] { (double) Math.random() * 1000,
+                                               (double) Math.random() * 1000 };
+        DataPoint dataPoint1 = new DataPoint(0, 0, coordinates1, new double[] { 0, 0 });
         assertTrue("dataPoint1.coordinates sollten {" + coordinates1[0]
                 + ", " + coordinates1[1] + "} sein", dataPoint1.coordinates.equals(coordinates1));
 
-        int[] coordinates2 = new int[] { (int) Math.random() * 1000, (int) Math.random() * 1000 };
+        double[] coordinates2 = new double[] { (double) Math.random() * 1000,
+                                               (double) Math.random() * 1000 };
         DataPoint dataPoint2 = new DataPoint(0, 0, coordinates2, 0.0f);
         assertTrue("dataPoint2.coordinates sollten {" + coordinates2[0]
                 + ", " + coordinates2[1] + "} sein", dataPoint2.coordinates.equals(coordinates2));
@@ -60,12 +62,14 @@ public class DataPointTest {
      */
     @Test
     public void rasterSizeShouldBeSetCorrectly() {
-        int[] rasterSize1 = new int[] { (int) Math.random() * 1000, (int) Math.random() * 1000 };
-        DataPoint dataPoint1 = new DataPoint(0, 0, rasterSize1, new int[] { 0, 0 });
+        double[] rasterSize1 = new double[] { (double) Math.random() * 1000,
+                                              (double) Math.random() * 1000 };
+        DataPoint dataPoint1 = new DataPoint(0, 0, rasterSize1, new double[] { 0, 0 });
         assertTrue("dataPoint1.rasterSize sollten {" + rasterSize1[0]
                 + ", " + rasterSize1[1] + "} sein", dataPoint1.coordinates.equals(rasterSize1));
 
-        int[] rasterSize2 = new int[] { (int) Math.random() * 1000, (int) Math.random() * 1000 };
+        double[] rasterSize2 = new double[] { (double) Math.random() * 1000,
+                                              (double) Math.random() * 1000 };
         DataPoint dataPoint2 = new DataPoint(0, 0, rasterSize2, 0.0f);
         assertTrue("dataPoint2.rasterSize sollten {" + rasterSize2[0]
                 + ", " + rasterSize2[1] + "} sein", dataPoint2.coordinates.equals(rasterSize2));
@@ -77,7 +81,7 @@ public class DataPointTest {
     @Test
     public void zoomLevelShouldBeSetCorrectly() {
         float zoomLevel1 = (float) Math.random() * 1000;
-        DataPoint dataPoint1 = new DataPoint(0, 0, new int[] { 0, 0 }, zoomLevel1);
+        DataPoint dataPoint1 = new DataPoint(0, 0, new double[] { 0, 0 }, zoomLevel1);
         assertEquals("dataPoint1.getZoomLevel() sollte " + zoomLevel1 + " sein",
                 zoomLevel1, dataPoint1.zoomLevel, 0.0001f);
     }
@@ -87,7 +91,7 @@ public class DataPointTest {
      */
     @Test
     public void toStringShouldStringGood() {
-        DataPoint dataPoint1 = new DataPoint(1, 2, new int[] { 3, 4 }, 42);
+        DataPoint dataPoint1 = new DataPoint(1, 2, new double[] { 3, 4 }, 42);
         assertEquals("""
                 DataPoint: {
                     dataId: 1,
@@ -97,7 +101,7 @@ public class DataPointTest {
                     zoomLevel: 42.0
                 }""", dataPoint1.toString());
 
-        DataPoint dataPoint2 = new DataPoint(5, 6, new int[] { 7, 8 }, new int[] { 9, 10 });
+        DataPoint dataPoint2 = new DataPoint(5, 6, new double[] { 7, 8 }, new double[] { 9, 10 });
         assertEquals("""
                 DataPoint: {
                     dataId: 5,
